pyproject.toml
[project]
name = "mcp-server-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.54.0",
    "arxiv>=2.2.0",
    "google-generativeai>=0.8.5",
    "langchain>=0.3.25",
    "langchain-openai>=0.3.22",
    "mcp-use>=1.3.0",
    "mcp[cli]>=1.9.3",
    "openai>=1.86.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
]
